dg import part 5 - watchdog/dotenv
Summary:
I think this is the last one for now.
