+++ python_modules/libraries/dagster-dg/dagster_dg/utils/filesystem.py
@@ -1,5 +1,4 @@
 import datetime
-import hashlib
 import time
 from collections.abc import Sequence
 from pathlib import Path
@@ -8,44 +7,8 @@ from typing import Any, Callable, Optional
 from watchdog.events import FileSystemEvent, FileSystemEventHandler
 from watchdog.observers import Observer
 
-from dagster_dg.utils import (
-    DEFAULT_FILE_EXCLUDE_PATTERNS,
-    clear_screen,
-    hash_directory_metadata,
-    hash_file_metadata,
-)
-
-
-def hash_paths(
-    paths: Sequence[Path],
-    includes: Optional[Sequence[str]] = None,
-    excludes: Sequence[str] = DEFAULT_FILE_EXCLUDE_PATTERNS,
-    error_on_missing: bool = True,
-) -> str:
-    """Hash the given paths, including their metadata.
-
-    Args:
-        paths: The paths to hash.
-        includes: A list of glob patterns to target, excluding files that don't match any of the patterns.
-        excludes: A list of glob patterns to exclude, including files that match any of the patterns. Defaults to
-            various Python-related files that are not relevant to the contents of the project.
-
-    Returns:
-        The hash of the paths.
-    """
-    hasher = hashlib.md5()
-    for path in paths:
-        if path.is_dir():
-            hash_directory_metadata(
-                hasher,
-                path,
-                includes=includes,
-                excludes=excludes,
-                error_on_missing=error_on_missing,
-            )
-        else:
-            hash_file_metadata(hasher, path, error_on_missing=error_on_missing)
-    return hasher.hexdigest()
+from dagster_dg.utils import DEFAULT_FILE_EXCLUDE_PATTERNS, clear_screen
+from dagster_dg.utils.paths import hash_paths
 
 
 class PathChangeHandler(FileSystemEventHandler):
