+++ python_modules/libraries/dagster-dg/dagster_dg_tests/cli_tests/import_perf_tests/test_import_perf.py
@@ -44,6 +44,8 @@ def test_import_perf():
         "http.server",
         "pydantic",
         "tomlkit",
+        "watchdog",
+        "dotenv",
     ]
     expensive_imports = [
         f"`{lib}`"
